pyproject.toml
[tool.poetry]
name = "curve-beancount"
version = "0.1.0"
description = ""
authors = ["Charlie Wren <cwren92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.37.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
pytest = "^7.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
